feat: finished

+++ order-service/src/main/java/com/example/org/entity/ProductView.java
@@ -20,7 +20,7 @@ public class ProductView {
     private Long id;
 
     @Column(name = "price")
-    private Long price;
+    private Integer price;
 
     @Column(name="name")
     private String name;
